tm 14 api connection done
