+++ app/src/main/java/pl/edu/pjatk/trainmate/api/RetrofitApiClient.java
@@ -3,7 +3,6 @@ package pl.edu.pjatk.trainmate.api;
 import static pl.edu.pjatk.trainmate.utils.Const.LOCAL_API_URL;
 
 import com.google.gson.GsonBuilder;
-
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
 
@@ -14,17 +13,17 @@ public class RetrofitApiClient {
     public static Retrofit getRetrofitInstance() {
         if (retrofit == null) {
             var gson = new GsonBuilder()
-                    .setLenient()
-                    .disableHtmlEscaping()
-                    .enableComplexMapKeySerialization()
-                    .serializeNulls()
-                    .enableComplexMapKeySerialization()
-                    .create();
+                .setLenient()
+                .disableHtmlEscaping()
+                .enableComplexMapKeySerialization()
+                .serializeNulls()
+                .enableComplexMapKeySerialization()
+                .create();
 
             retrofit = new Retrofit.Builder()
-                    .baseUrl(LOCAL_API_URL)
-                    .addConverterFactory(GsonConverterFactory.create(gson))
-                    .build();
+                .baseUrl(LOCAL_API_URL)
+                .addConverterFactory(GsonConverterFactory.create(gson))
+                .build();
         }
 
         return retrofit;
