+++ app/src/main/java/pl/edu/pjatk/trainmate/api/plan/Plan.java
@@ -1,10 +1,8 @@
 package pl.edu.pjatk.trainmate.api.plan;
 
 import com.google.gson.annotations.SerializedName;
-
 import java.util.ArrayList;
 import java.util.List;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
