+++ app/src/main/java/pl/edu/pjatk/trainmate/api/plan/TrainingUnitProjection.java
@@ -1,11 +1,9 @@
 package pl.edu.pjatk.trainmate.api.plan;
 
 import com.google.gson.annotations.SerializedName;
-
 import java.time.DayOfWeek;
 import java.util.ArrayList;
 import java.util.List;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
