+++ app/src/main/java/pl/edu/pjatk/trainmate/utils/Const.java
@@ -2,9 +2,6 @@ package pl.edu.pjatk.trainmate.utils;
 
 public class Const {
 
-    //TODO: WYJEBAC to
-    public static String REFRESH_TOKEN;
-    public static String TOKEN_TEST;
     public static final String REFRESH_ACTIVE = "refreshActive";
     public static final String PREFS_NAME = "preferences";
     public static final String PREF_UNAME = "Username";
@@ -14,9 +11,12 @@ public class Const {
     public static final String CLIENT_ID = "train-mate";
     public static final String DEFAULT_STRING_VALUE = "";
     public static final String LOCAL_KEYCLOAK_URL = "http://10.0.2.2:8123";
-    public static final String LOCAL_API_URL = "http://10.0.2.2:8080/api/tm-core/";
+    public static final String LOCAL_API_URL = "http://10.0.2.2:8081/api/tm-core/";
     public static final String TOKEN_REFRESH_GRANT_TYPE = "refresh_token";
     public static final String REFRESH_TAG = "REFRESH TOKEN";
     public static final String REFRESH_TOKEN_FAIL = "Token refresh malfunction";
     public static final String LOGIN_FAIL_ANNOUNCEMENT = "Incorrect login or password";
+    public static final String TOKEN_TYPE = "Bearer ";
+    public static final String API_FAIL = "Data downloading from API malfunction";
+    public static final String API_TAG = "API FAIL";
 }
