+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -13,13 +13,13 @@
 
     <fragment
       android:id="@+id/nav_plan"
-      android:name="pl.edu.pjatk.trainmate.ui.report.ReportFragment"
+      android:name="pl.edu.pjatk.trainmate.ui.plan.PlanFragment"
       android:label="@string/menu_plan"
       tools:layout="@layout/fragment_report" />
 
     <fragment
       android:id="@+id/nav_report"
-      android:name="pl.edu.pjatk.trainmate.ui.plan.PlanFragment"
+      android:name="pl.edu.pjatk.trainmate.ui.report.ReportFragment"
       android:label="@string/menu_report"
       tools:layout="@layout/fragment_plan" />
 </navigation>
